requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Compress==1.15
Flask-SQLAlchemy>=3.1.1
yt-dlp>=2024.8.6
requests>=2.32.2
urllib3>=2.1.0
curl-cffi>=0.13.0
instaloader>=4.10.3
fake-useragent>=1.4.0

# Production & Security
python-dotenv>=1.0.0
Flask-Limiter>=3.5.0
Flask-Babel>=4.0.0
gunicorn>=21.2.0

# PDF Conversion Libraries
pdf2docx>=0.5.6
PyPDF2>=3.0.0
pdf2image>=1.16.3
img2pdf>=0.5.0
python-docx>=1.1.0
docx2pdf>=0.1.8
pillow>=10.0.0

# Image Processing Libraries
Pillow>=10.0.0
opencv-python>=4.8.0
rembg>=2.0.50
numpy>=1.24.0

# Excel/Office Support
openpyxl>=3.1.2
python-pptx>=0.6.21
tabula-py>=2.8.0
reportlab>=4.0.0

# PDF Processing (Advanced)
PyMuPDF>=1.23.0  # Also known as 'fitz', for PDF rendering

# File Validation
python-magic>=0.4.27
python-magic-bin>=0.4.14  # For Windows

# Utilities
Werkzeug>=3.0.0
